requirements.txt
xgboost>=1.7.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
mlflow>=2.0.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0